// nextjs/brisk-web-ssr/src/api/welfare/types.gen.ts
export type WelfareReceiveRecordResponse = {
    /**
     * 币种
     */
    currency?: string;
    /**
     * 过期时间
     */
    expiredAt?: number;
    /**
     * 主键id
     */
    id?: string;
    /**
     * 索引
     */
    index?: number;
    /**
     * 是否手气最佳
     */
    isLucky?: boolean;
    /**
     * 是否新用户
     */
    isNewUser?: boolean;
    /**
     * 当前数据是否是自己的领取记录：是-true，否-false
     */
    isSelf?: boolean;
    /**
     * 领取人昵称
     */
    nickName?: string;
    /**
     * 已领取金额
     */
    receiveAmount?: string;
    /**
     * 领取时间
     */
    receiveAt?: string;
    /**
     * 领取状态  0-待领取,1-已领取,2-入账成功,11-锁定,12-退回
     */
    receiveStatus?: number;
    /**
     * 剩余时间
     */
    remainingAt?: number;
    /**
     * 发送金额
     */
    sendAmount?: string;
    /**
     * 发送人昵称
     */
    sendNickName?: string;
    /**
     * 福利状态：-1-异常状态,0-不存在，1-领取中，2-抢完，3-过期
     */
    welfareStatus?: number;
};

export type WelfareReceiveRecordV2Response = {
    /**
     * 币种
     */
    currency?: string;
    /**
     * 过期时间
     */
    expiredAt?: number;
    /**
     * 主键id
     */
    id?: string;
    /**
     * 索引
     */
    index?: number;
    /**
     * 是否手气最佳
     */
    isLucky?: boolean;
    /**
     * 是否新用户
     */
    isNewUser?: boolean;
    /**
     * 当前数据是否是自己的领取记录：是-true，否-false
     */
    isSelf?: boolean;
    /**
     * 领取人昵称
     */
    nickName?: string;
    /**
     * 已领取金额
     */
    receiveAmount?: string;
    /**
     * 领取时间
     */
    receiveAt?: string;
    /**
     * 领取状态  0-待领取,1-已领取,2-入账成功,11-锁定,12-退回
     */
    receiveStatus?: number;
    /**
     * 剩余时间
     */
    remainingAt?: number;
    /**
     * 发送金额
     */
    sendAmount?: string;
    /**
     * 发送人昵称
     */
    sendNickName?: string;
    /**
     * 福利状态：-1-异常状态,0-不存在，1-领取中，2-抢完，3-过期
     */
    welfareStatus?: number;
};

export type WelfareReceiveV2Request = {
    /**
     * 渠道
     */
    channel?: 'APP' | 'M' | 'WEB';
    /**
     * 福利口令code
     */
    code?: string;
    /**
     * 国际区号
     */
    countryCode?: string;
    /**
     * 邮箱
     */
    email?: string;
    /**
     * 领取人的电话
     */
    phone?: string;
};

export type WelfareRecordExistsResponse = {
    /**
     * 祝福语
     */
    bestWishes?: string;
    /**
     * 币种
     */
    currency?: string;
    /**
     * 红包封面地址
     */
    extendUrl?: string;
    /**
     * 是否是新人红包
     */
    isNewBee?: boolean;
    /**
     * 是否已领取：false-未领取，true-已领取
     */
    isReceive?: boolean;
    /**
     * 福利发送人
     */
    nickName?: string;
    /**
     * 已经领取的金额
     */
    receiveAmount?: string;
    /**
     * 领取状态  0-待领取,1-已领取,2-入账成功,11-锁定,12-退回
     */
    receiveStatus?: number;
    /**
     * 发送福利表id
     */
    sendRecordId?: string;
    /**
     * 福利状态：-1-异常状态,0-不存在，1-领取中，2-抢完，3-过期
     */
    welfareStatus?: number;
    /**
     * 福利类型类型：0-实时福利(拼)、1-实时福利(固定金额）
     */
    welfareType?: number;
};

export type WelfareResponse = {
    /**
     * 祝福语
     */
    bestWishes?: string;
    /**
     * 发送币种
     */
    currency?: string;
    /**
     * 红包封面地址
     */
    extendUrl?: string;
    /**
     * 首次领取时间
     */
    firstAt?: string;
    /**
     * 主键id
     */
    id?: string;
    /**
     * 是否是新人红包
     */
    isNewBee?: boolean;
    /**
     * 当前登录用户是否已领取：false-未领取，true-已领取
     */
    isReceive?: boolean;
    /**
     * 当前福利是否是自己发送的：是-true，否-false
     */
    isSelf?: boolean;
    /**
     * 最后领取时间
     */
    lastAt?: string;
    /**
     * 福利发送人昵称
     */
    nickName?: string;
    receive?: WelfareReceiveRecordResponse;
    /**
     * 已领取币种数量
     */
    receiveAmount?: string;
    /**
     * 已领取个数
     */
    receiveNum?: number;
    /**
     * 发送币种数量
     */
    sendAmount?: string;
    /**
     * 红包发送时间
     */
    sendAt?: string;
    /**
     * 发送个数
     */
    sendNum?: number;
    /**
     * 福利状态：-1-异常状态,0-不存在，1-领取中，2-抢完，3-过期
     */
    welfareStatus?: number;
    /**
     * 获取福利类型：0-实时福利(拼)、1-实时福利(普通）
     */
    welfareType?: number;
};